⏪ Revert move of  hf_hub_url method
