+++ src/huggingface_hub/utils/endpoint_helpers.py
@@ -17,98 +17,7 @@ with the aim for a user-friendly interface
 import math
 import re
 from dataclasses import dataclass
-from typing import List, Optional, Union
-from urllib.parse import quote
-
-from ..constants import (
-    DEFAULT_REVISION,
-    HUGGINGFACE_CO_URL_TEMPLATE,
-    REPO_TYPES,
-    REPO_TYPES_URL_PREFIXES,
-)
-
-
-def hf_hub_url(
-    repo_id: str,
-    filename: str,
-    *,
-    subfolder: Optional[str] = None,
-    repo_type: Optional[str] = None,
-    revision: Optional[str] = None,
-) -> str:
-    """Construct the URL of a file from the given information.
-
-    The resolved address can either be a huggingface.co-hosted url, or a link to
-    Cloudfront (a Content Delivery Network, or CDN) for large files which are
-    more than a few MBs.
-
-    Args:
-        repo_id (`str`):
-            A namespace (user or an organization) name and a repo name separated
-            by a `/`.
-        filename (`str`):
-            The name of the file in the repo.
-        subfolder (`str`, *optional*):
-            An optional value corresponding to a folder inside the repo.
-        repo_type (`str`, *optional*):
-            Set to `"dataset"` or `"space"` if uploading to a dataset or space,
-            `None` or `"model"` if uploading to a model. Default is `None`.
-        revision (`str`, *optional*):
-            An optional Git revision id which can be a branch name, a tag, or a
-            commit hash.
-
-    Example:
-
-    ```python
-    >>> from huggingface_hub import hf_hub_url
-
-    >>> hf_hub_url(
-    ...     repo_id="julien-c/EsperBERTo-small", filename="pytorch_model.bin"
-    ... )
-    'https://huggingface.co/julien-c/EsperBERTo-small/resolve/main/pytorch_model.bin'
-    ```
-
-    <Tip>
-
-    Notes:
-
-        Cloudfront is replicated over the globe so downloads are way faster for
-        the end user (and it also lowers our bandwidth costs).
-
-        Cloudfront aggressively caches files by default (default TTL is 24
-        hours), however this is not an issue here because we implement a
-        git-based versioning system on huggingface.co, which means that we store
-        the files on S3/Cloudfront in a content-addressable way (i.e., the file
-        name is its hash). Using content-addressable filenames means cache can't
-        ever be stale.
-
-        In terms of client-side caching from this library, we base our caching
-        on the objects' entity tag (`ETag`), which is an identifier of a
-        specific version of a resource [1]_. An object's ETag is: its git-sha1
-        if stored in git, or its sha256 if stored in git-lfs.
-
-    </Tip>
-
-    References:
-
-    -  [1] https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/ETag
-    """
-    if subfolder is not None:
-        filename = f"{subfolder}/{filename}"
-
-    if repo_type not in REPO_TYPES:
-        raise ValueError("Invalid repo type")
-
-    if repo_type in REPO_TYPES_URL_PREFIXES:
-        repo_id = REPO_TYPES_URL_PREFIXES[repo_type] + repo_id
-
-    if revision is None:
-        revision = DEFAULT_REVISION
-    return HUGGINGFACE_CO_URL_TEMPLATE.format(
-        repo_id=repo_id,
-        revision=quote(revision, safe=""),
-        filename=filename,
-    )
+from typing import List, Union
 
 
 def _filter_emissions(
